requirement.txt
appdirs==1.4.3
attrs==21.2.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
comet-ml==3.9.1
configobj==5.0.6
contextlib2==0.6.0
cycler==0.10.0
distlib==0.3.0
distro==1.4.0
dulwich==0.20.21
everett==1.0.3
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
jsonschema==3.2.0
kiwisolver==1.3.1
lockfile==0.12.2
matplotlib==3.4.2
msgpack==0.6.2
numpy==1.20.2
nvidia-ml-py3==7.352.0
opt-einsum==3.3.0
packaging==20.3
pep517==0.8.2
Pillow==8.2.0
progress==1.5
pyparsing==2.4.6
pyro-api==0.1.2
pyro-ppl==1.3.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytoml==0.1.21
requests==2.22.0
requests-toolbelt==0.9.1
retrying==1.3.3
six==1.14.0
torch==1.8.1
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.25.8
webencodings==0.5.1
websocket-client==0.59.0
wrapt==1.12.1
wurlitzer==2.1.0
